+++ src/main/java/codesquad/webserver/file/FileHttpResponseCreator.java
@@ -1,6 +1,5 @@
-package codesquad.webserver.handler;
+package codesquad.webserver.file;
 
-import codesquad.webserver.file.StaticFileReader;
 import codesquad.webserver.http.HttpResponse;
 import codesquad.webserver.http.type.ContentType;
 import codesquad.webserver.http.type.HttpHeader;
@@ -15,6 +14,11 @@ public class FileHttpResponseCreator {
 
     private FileHttpResponseCreator() {}
 
+    /**
+     * 전달된 경로를 기반으로 정적 파일 응답을 생성합니다.
+     * @param resourcePath 읽어들일 파일 경로
+     * @return static 패키지 기준으로 불러들인 파일
+     */
     public static HttpResponse create(String resourcePath) {
         logger.debug("정적 파일 응답을 생성합니다. Resource: {}", resourcePath);
         String mimeType = getMimeType(resourcePath);
@@ -25,6 +29,12 @@ public class FileHttpResponseCreator {
         return HttpResponse.ok(headers, fileData);
     }
 
+    /**
+     * 전달된 경로를 기반으로 파일을 읽고 템플릿 엔진을 통해 동적 응답을 생성합니다.
+     * @param resourcePath 읽어들일 파일 경로
+     * @param templateData 템플릿 엔진을 통해 치환할 데이터
+     * @return static 패키지 기준으로 불러들인후 템플릿 엔진을 통해 생성된 파일
+     */
     public static HttpResponse create(String resourcePath, Map<String, String> templateData) {
         logger.debug("동적 파일 응답을 생성합니다. Resource: {}, MappingData: {}", resourcePath, templateData);
         String mimeType = getMimeType(resourcePath);
@@ -42,12 +52,7 @@ public class FileHttpResponseCreator {
         return ContentType.from(extension).getMimeType();
     }
 
-    /**
-     *
-     * @param path
-     * @return
-     * @throws IllegalArgumentException 파일 읽기 실패시 발생
-     */
+    // @throws IllegalArgumentException 파일 읽기 실패시 발생
     private static String getStaticFile(String path) {
         return staticFileReader.read(path);
     }
